+++ js/data-test.js
@@ -1,0 +1,48 @@
+const DESCRIPTIONS = [
+  'Фография интересная',
+  'Фография красивая',
+  'Фография черно-белая',
+  'Фография цветная',
+  'Фография ужасная',
+  'Фография печальная',
+  'Фография веселая'
+];
+
+const NAMES = [
+  'Артем',
+  'Светлана',
+  'Игорь',
+  'Мария',
+  'Ксения',
+  'Петр',
+  'Иван',
+  'Владимир',
+  'Алексей',
+  'Ассоль',
+  'Лилия',
+  'Максим',
+  'Константин',
+  'Евгения',
+  'Михаил',
+  'Станислав',
+  'Елена',
+  'Александр',
+  'Марина',
+  'Вероника',
+  'Бен',
+  'Стен',
+  'Глен',
+  'Ален',
+  'Мирослава'
+];
+
+const SENTENCES = [
+  'Всё отлично! В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export {DESCRIPTIONS,NAMES,SENTENCES};
